refactor: restructuring convex files
