+++ convex/users/_helpers/getCurrentUser.ts
@@ -1,9 +1,11 @@
 import { QueryCtx } from "@convex/_generated/server";
 
-import { userByExternalId } from "./user_by_external_id";
+import { userByExternalId } from "./userByExternalId";
 
 export async function getCurrentUser(ctx: QueryCtx) {
   const identity = await ctx.auth.getUserIdentity();
+
   if (!identity) return null;
+
   return userByExternalId(ctx, identity.subject);
 }
