+++ convex/users/_helpers/index.ts
@@ -1,3 +1,3 @@
-export * from "./get_current_user";
-export * from "./upsert_user_record";
-export * from "./user_by_external_id";
+export * from "./getCurrentUser";
+export * from "./upsertUserRecord";
+export * from "./userByExternalId";
