+++ convex/users/_helpers/userByExternalId.ts
