--- convex/users/_mutations/index.ts
@@ -1,1 +1,0 @@
-export * from "./upsertUser";
