--- convex/users/mutations.ts
@@ -1,1 +1,0 @@
-export * from "./_mutations";
