+++ convex/users/mutations/upsertUser.test.ts
@@ -2,12 +2,13 @@ import { convexTest } from "convex-test";
 import { describe, expect, test } from "vitest";
 
 import { api } from "../../_generated/api";
+import { modules } from "../../_testModule";
 import schema from "../../schema";
 
 describe("upsertUser", () => {
   test("throws when not authenticated", async () => {
     const t = convexTest(schema, modules);
-    await expect(t.mutation(api.users.mutations.upsertUser, {})).rejects.toThrow(
+    await expect(t.mutation(api.users.mutations.upsertUser.mutate, {})).rejects.toThrow(
       "Not authenticated",
     );
   });
@@ -20,22 +21,22 @@ describe("upsertUser", () => {
       givenName: "John",
       familyName: "Doe",
     });
-    const userId = await asUser.mutation(api.users.mutations.upsertUser, {});
+    const userId = await asUser.mutation(api.users.mutations.upsertUser.mutate, {});
     expect(userId).toBeDefined();
   });
 
   test("returns existing user id on repeat call", async () => {
     const t = convexTest(schema, modules);
     const asUser = t.withIdentity({ subject: "clerk_abc123", email: "test@example.com" });
-    const firstId = await asUser.mutation(api.users.mutations.upsertUser, {});
-    const secondId = await asUser.mutation(api.users.mutations.upsertUser, {});
+    const firstId = await asUser.mutation(api.users.mutations.upsertUser.mutate, {});
+    const secondId = await asUser.mutation(api.users.mutations.upsertUser.mutate, {});
     expect(firstId).toEqual(secondId);
   });
 
   test("sets hasCompletedOnboarding to false on creation", async () => {
     const t = convexTest(schema, modules);
     const asUser = t.withIdentity({ subject: "clerk_new", email: "new@example.com" });
-    await asUser.mutation(api.users.mutations.upsertUser, {});
+    await asUser.mutation(api.users.mutations.upsertUser.mutate, {});
     const user = await t.run((ctx) =>
       ctx.db
         .query("users")
@@ -48,7 +49,7 @@ describe("upsertUser", () => {
   test("sets isPublic to false on creation", async () => {
     const t = convexTest(schema, modules);
     const asUser = t.withIdentity({ subject: "clerk_pub", email: "pub@example.com" });
-    await asUser.mutation(api.users.mutations.upsertUser, {});
+    await asUser.mutation(api.users.mutations.upsertUser.mutate, {});
     const user = await t.run((ctx) =>
       ctx.db
         .query("users")
@@ -66,7 +67,7 @@ describe("upsertUser", () => {
       givenName: "Alice",
       familyName: "Smith",
     });
-    await asUser.mutation(api.users.mutations.upsertUser, {});
+    await asUser.mutation(api.users.mutations.upsertUser.mutate, {});
     const user = await t.run((ctx) =>
       ctx.db
         .query("users")
@@ -77,5 +78,3 @@ describe("upsertUser", () => {
     expect(user?.lastName).toBe("Smith");
   });
 });
-
-const modules = import.meta.glob("./**/*.ts");
