+++ convex/users/mutations/upsertUser.ts
@@ -1,7 +1,7 @@
 import { mutation } from "../../_generated/server";
 import { upsertUserRecord } from "../_helpers";
 
-export const upsertUser = mutation({
+export const mutate = mutation({
   args: {},
   handler: async (ctx) => {
     const identity = await ctx.auth.getUserIdentity();
